+++ app/src/main/java/io/ymsoft/easypick/features/domain/model/Candidate.kt
@@ -1,8 +1,6 @@
 package io.ymsoft.easypick.features.domain.model
 
-import androidx.room.Entity
-import androidx.room.ForeignKey
-import androidx.room.PrimaryKey
+import androidx.room.*
 import java.util.*
 
 @Entity(
@@ -17,7 +15,7 @@ data class Candidate(
     var name: String,
     var content: String = "",
     val created: Long = Date().time,
-    val groupId: Int? = null,
+    val groupId: Long? = null,
     @PrimaryKey(autoGenerate = true) val id: Int? = null
 )
